Refactor summary method + correct calculation for response rate
Existing methods look at the completion status of the response. A response
is complete only if it has answers for all questions. Really we want a
per-question response rate, which is the number of Answers for that question
divided by the number of expected answers for that question (i.e., the number
of Responses for the pollrun).