Update text on flows page
On the page at "/poll/":

New text:
Change the wording of "Polls" -> "Flows" on this page

New description:
"Latest statistics on flows that are being tracked in TracPro. Click on the 'Select Flows' button to change the flows you are tracking. All flow data is tracked going forward. If you need to track flow data for dates in the past, ask your administrator to fetch runs for those dates."