Redirect chooser when org known
This fixes a longtime annoyance.

If you visit the base domain of tracpro (without the organization
prefix), you see the chooser view telling you that you need
to pick an organization.

My natural instinct is to edit the URL to stick the organization
subdomain and a dot in front of the URL and hit Enter to reload.

But before this fix, that loaded the chooser view again, which
again displayed the message that you needed to pick an
organization. At this point, that's just wrong.

This change will redirect to the site home page from the
chooser if the organization is known.