[TD-58] Let orgs optionally sum multiple responses
More specifically, let orgs configure an optional
alternate behavior when the same contact answers the
same numeric question more than once on the same day.
Ordinarily we use the last answer submitted, but they
can instead choose to sum the numeric responses.

This is an org-wide option, set in the Org's configuration
page that is only accessible to org administrators
(or maybe to superusers? something like that).