Fix bug syncing twitter contacts
RapidPro seems to be converting URNs that use
Twitter handles to ones using Twitter IDs, which tracpro
doesn't know about. Update the list of URN schemes that
we understand to the current list on RapidPro.  (All we
really need to know about these is a translatable label
for each one, no additional code needed on our end.)