When a contact is edited on Tracpro, push the changes to Rapidpro
Current behavior is to try to merge the data on Tracpro with the
data on Rapidpro, but that leads to unwanted behavior in some cases.
For example, if you remove a contact from a group on Tracpro, it'll
still belong on Rapidpro, and when the groups are merged, the
Tracpro contact will end up back in the group again when the user
wanted them removed from it.

This also allows removing some code that's no longer needed, which
is always nice.