Sync contacts that are no longer in groups of interest
If a contact was edited on RapidPro and removed from
all the groups our org was following, we'd never see that
because we only fetch contacts in our groups. This change
will mark inactive any contacts the org currently has
that were not returned by RapidPro, since we know they're
not in any group we're interested in anymore.