A faster way to handle sums of answers
Instead of going through all the org's answers and updating them in
the database when someone changes the setting, compute the value we
need at runtime. It probably slows the pages down some, but changing
the value was taking more than 30 minutes and then timing out trying
to update all the answers.